refactor(payment): Phase 8 relocate InternalOutboxEndpoints + Program.cs manifest
- Move Endpoints/InternalOutboxEndpoints.cs to Infrastructure/Outbox/
  with namespace Payment.Service.Infrastructure.Outbox; delete the
  now-empty Endpoints/ folder. Public extension name preserved
  (RegisterInternalOutboxEndpoints), RequireService gate preserved.
- Drop the obsolete `using Payment.Service.Endpoints;` from Program.cs.
  Program.cs already reads as a slice manifest after Phase 7d (no per-
  handler AddScoped, no per-event AddEventHandler — all in slice
  extensions); the only top-level AddScoped registrations remaining
  are infrastructure seams (MessageCorrelationContext +
  DomainEventOutboxInterceptor) called out as acceptable in the issue.
- Pre-existing MessagingProviderBootTests RabbitMQ failure unchanged
  (eager-broker defect, blocked on shared-lib bump).
